prog10/pyproject.toml
[tool.poetry]
name = "prog10"
version = "0.1.0"
description = ""
authors = ["JadwigaS <80515572+JadwigaS@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
bs4 = "^0.0.2"
pillow = "^10.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
